requirements.txt
openai
pillow
requests
ebooklib
python-dotenv
replicate
groq
torch
transformers

